+++ events/onNewDevice.js
@@ -3,7 +3,11 @@ const _ = require('lodash');
 const fs = require('fs');
 const path = require('path');
 
-module.exports = function onNewDeviceFn(knownDevices, id, topic, addOn, topicNotify, deviceAWS) {
+const AddOnHandler = require('../AddOnHandler');
+
+module.exports = function onNewDeviceFn(knownDevices, topicMess, topicNotify, deviceAWS) {
+  const { topic, deviceType, name } = topicMess;
+  const id = topicMess.idIn;
   return function onNewDevice(newDevice) {
     const payload = {
       status: {
@@ -15,27 +19,38 @@ module.exports = function onNewDeviceFn(knownDevices, id, topic, addOn, topicNot
     debug(JSON.stringify(payload));
     deviceAWS.publish(topicNotify, JSON.stringify(payload));
 
-    const { protocol } = newDevice;
+    const protocol = deviceType.protocols[0];
+
     const devices = knownDevices[protocol];
     _.unset(newDevice, 'protocol');
     const device = {
       id,
       topic,
-      addOn,
+      deviceType,
+      name,
       ...newDevice,
     };
     if (!devices) {
       _.set(knownDevices, protocol, [device]);
+      const addOnHandler = AddOnHandler.getInstance();
+      const instance = addOnHandler.get(device.id);
+      instance.setKnownDevices(knownDevices[protocol]);
     } else {
       devices.push(device);
+      devices.forEach((deviceLoop) => {
+        const deviceId = deviceLoop.id;
+        const addOnHandler = AddOnHandler.getInstance();
+        const instance = addOnHandler.get(deviceId);
+        instance.setKnownDevices(knownDevices[protocol]);
+      });
     }
-    const configPath = path.resolve(__dirname, '..', 'knownDevices.config');
+
+    const configPath = process.env.IS_HOST ? '/home/pi/service/knownDevices.config' : '/home/node/gate/service/knownDevices.config';
     return fs.writeFile(configPath, JSON.stringify(knownDevices), (err) => {
       if (err) {
         debug(err);
       }
       debug('File written successfully');
     });
-    return;
   };
 };