+++ pair.js
@@ -2,16 +2,19 @@ const debug = require('debug')('gate');
 const initAddOn = require('./subdevice/initAddOn');
 const publishData = require('./subdevice/publishData');
 const onNewDeviceFn = require('./events/onNewDevice');
+const onStatusFn = require('./events/onStatus');
+const onPairingTimeoutFn = require('./events/onPairingTimeout');
+const onRemovedDeviceFn = require('./events/onRemovedDevice');
 
 module.exports = function pairSubdevice(topicNotify, device, topicMess, knownDevices) {
   const responseTopic = topicNotify.replace('/post', '');
+  const { topic, deviceType, idIn } = topicMess;
+  const onNewDevice = onNewDeviceFn(knownDevices, topicMess, responseTopic, device);
+  const onRemovedDevice = onRemovedDeviceFn(knownDevices, deviceType, responseTopic, device);
+  const onPairingTimeout = onPairingTimeoutFn(responseTopic, device);
+  const onStatus = onStatusFn(device);
 
-  const { topic } = topicMess;
-  const { idIn } = topicMess;
-  // const { deviceType } = topicMess;
-  const { addOn } = topicMess;
-  const onNewDevice = onNewDeviceFn(knownDevices, idIn, topic, addOn, responseTopic, device);
-  return initAddOn(addOn, knownDevices, (err, addOnInstance) => {
+  return initAddOn(idIn, deviceType, knownDevices, (err, addOnInstance) => {
     addOnInstance.init();
     debug('start discovering device...');
 
@@ -33,5 +36,9 @@ module.exports = function pairSubdevice(topicNotify, device, topicMess, knownDev
       });
     });
     addOnInstance.on('newDevice', onNewDevice);
+    addOnInstance.on('deviceRemoved', onRemovedDevice);
+    addOnInstance.on('status', onStatus);
+    addOnInstance.on('timeoutDiscovering', onPairingTimeout);
+
   });
 };