+++ subdevice/initAddOn.js
@@ -1,23 +1,32 @@
-const debug = require('debug');
+const debug = require('debug')('gate');
 const installAddOn = require('./installAddOn');
+const AddOnHandler = require('../AddOnHandler');
 
-module.exports = function initAddOn(addOn, knownDevices, callback) {
+module.exports = function initAddOn(deviceId, deviceType, knownDevices, callback) {
+  const addOnHandler = AddOnHandler.getInstance();
+  const { addOn, pairingMethods } = deviceType;
   let GateAddOnClass;
   try {
     GateAddOnClass = require(addOn);
   } catch (error) {
-    console.log('ERRRR', error);
-    return installAddOn(addOn, knownDevices, (err, addOnInstance) => {
+    debug('installing addOn');
+    return installAddOn(deviceId, deviceType, knownDevices, (err, addOnInstance) => {
       if (err) {
-        console.log(err);
+        debug(err);
       }
+      addOnHandler.add(addOnInstance);
       return callback(null, addOnInstance);
     });
   }
-  console.log('known devi', knownDevices);
-  const addOnInstance = new GateAddOnClass(knownDevices);
-  console.log('created device instance');
+  let options = {};
+  if (pairingMethods) {
+    options = {
+      touchlink: pairingMethods.indexOf('touchlink') !== -1,
+    };
+  }
+  const addOnInstance = new GateAddOnClass(deviceId, deviceType, knownDevices, options);
+  debug('created device instance');
+
+  addOnHandler.add(addOnInstance);
   return callback(null, addOnInstance);
-
 };
-