+++ CHANGELOG.md
@@ -6,6 +6,40 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.2.4] - 2020-03-18
+### Added
+- README.md
+
+## [1.2.3] - 2020-03-18
+### Changed
+- Bumped Node.js version 12.16.1
+- Fixed Zigbee protocol script
+
+## [1.2.2] - 2020-03-17
+### Changed
+- Bumped dependencies
+
+## [1.2.1] - 2020-03-04
+### Added
+- PairingIndicatorsSubdevice component
+- Unlink feature
+
+## [1.2.0] - 2020-03-02
+### Added
+- first test for `AddOnHandler`
+- support for `control/get` mqtt messages
+
+## [1.1.2] - 2020-02-26
+### Added
+- `master` branch added to build.yml 
+
+## [1.1.1] - 2020-02-26
+### Changed
+- No changes, waiting for build
+
+## [1.1.0] - 2020-02-25
+### Added
+- add feature to control devices
 
 ## [1.0.1] - 2020-02-17
 ### Added